Use NULL for validator columns; add upgrader.
V1 of chaind used '-1' to mark a validator epoch column as
FAR_FUTURE_EPOCH.  The actual value of FAR_FUTURE_EPOCH is not
compatible with the PostgreSQL 'BIGINT' column, as it is outside of its
range.  Although -1 worked, it made querying the database harder than it
needed to be with multiple constraints, and could cause confusion with
differing definitions of the conditions for states.

This change alters the schema of t_validators to allow NULL values in
the relevant columns, and updates existing values of FAR_FUTURE_EPOCH to
be NULL.

This change also introduces a generic database upgrader, which is used
to carry out the required changes to the database schema and data.  The
upgrader can be used for additional changes in the schema as V2
progresses.