Make log levels hierarchical.
This change makes configuration of log levels using the `log-level`
configuration parameters hierarchical.  Given a path in dot notation,
for example a.b, the following configuration variables will be searched
for a log value:

  - a.b.log-level
  - a.log-level
  - log-level

The first variable containing a value will be used.

This allows for simple setting of the global level with individual as
well as group overrides.