Add vote finalizer.
This change adds the fields f_target_correct and f_head_correct to the
t_attestations table.  It also adds code to the finalizer that augments
attestations with details of the correctness of their target and head
votes.  It attempts to do this as finality information is obtained, as
well as on start-up.

Because modules in chaind are loosely coupled, it is possible for the
finalizer to not have the required information in the chaind database to
carry out its operations.  In these situations it will retry
incrementally at each new finality checkpoint, as for all indeterminate
attestations on start-up.