Add finalizer.
This adds the summarizer module to chaind.  The summarizer module
provides summary information at the epoch, block and validator levels,
allowing for many common queries to be run much more quickly and to
allow users to query this type of data without requiring the in-depth
knowledge of the internals of the beacon chain required to calculate
them from the component tables.

Becaues the summarizer summarizes information obtained by other modules
this requires some tweaks to the validators and finalizer modules to
ensure the correct information is present before the summarizer runs.