Fail fast if data unavailable.
This patch changes the behaviour of chaind in the face of a failure to
obtain data from "mark as missed and carry on" to "stop this attempt at
data collection".  With the subsequent introduction of the finalizer and
summarizer the former behaviour can result in incorrect summaries,
whereas the latter can result in delayed summaries.  The latter is
preferable.